shared global prisma for hub
