+++ internal-packages/hub-db/src/index.ts
@@ -1,1 +1,2 @@
 export * from "../generated/index.js";
+export { getPrismaClient } from "./client.js";
